docs: update AUTHORS format (#172)
Addressing an email received after the second submission of prqlr 0.5.2
(171).

> Thanks, we see this still has things like
>
>
>   - either (version 1.9.0): bluss
>
> Who or what is 'bluss'?
>
> - wasi (version 0.11.0+wasi-snapshot-preview1): The Cranelift Project
Developers
>
>
> What is that? Seems a project that has changed its name recently, so
needs clarification.
>
> We really need to see the authors and copyright holders.
> As you include their code, they hold partial copyright on your
package. Even if the rust people seem to be relaxed, we want to know who
holds copyright on the software we distribute. And we believe one has to
decalre this cleanly.

Lists the owner of each crate registered on Crates.io.